pyproject.toml
[project]
name = "ai-solution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-tinymce>=4.1.0",
    "django>=5.1.4",
    "pillow>=11.0.0",
    "pytz>=2024.2",
    "scikit-learn>=1.6.0",
    "numpy>=2.2.1",
]
